adapters/projects/logx/deposit_test.go
package logx_test

import (
	"context"
	"testing"

	"github.com/ethereum/go-ethereum/common"
	"github.com/ethereum/go-ethereum/ethclient"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"
	"github.com/taikoxyz/trailblazer-adapters/adapters"
	"github.com/taikoxyz/trailblazer-adapters/adapters/projects/logx"
)

func TestTransferRemoteIndexer(t *testing.T) {
	taikoRPC := "https://rpc.taiko.xyz"
	blocknumber := int64(448072)

	ctx := context.Background()

	client, err := ethclient.Dial(taikoRPC)
	require.NoError(t, err)

	// Use the new TransferRemoteIndexer instead of DepositIndexer
	indexer := logx.NewTransferRemoteIndexer(client, []common.Address{common.HexToAddress(logx.DepostiAddress)})

	logs, err := adapters.GetLogs(ctx, client, indexer.Addresses(), blocknumber)
	require.NoError(t, err)

	// Call the Index function and check results for the SentTransferRemote event
	whitelist, err := indexer.Index(ctx, logs...)
	assert.NoError(t, err)
	assert.Len(t, whitelist, 8)  // Adjust this as needed based on the expected event count
}
